refactor(debug): extract the debug/diagnostics/logs router — pure move
Phase 2 Task 8 (debug slice), step 2. Moves twelve route-methods out of
src/api_server.py into src/debug/routes.py behind APIRouter(tags=["debug"]):

  POST /debug/blank            POST /debug/clear-cache
  POST /debug/fill             GET  /debug/cache-status
  POST /debug/info             GET  /debug/system-info
  POST /debug/test-connection  GET  /debug/network-diagnostics
  GET  /cache-status           POST /clear-cache
  POST /force-refresh          GET  /logs

The four siblings move with the /debug/* eight because they are the same
subsystem — the same board client, the same caches, the same support flow —
and splitting them would leave the domain half-converted.

Pure move, so the safety gate holds:

* tests/golden/api_routes.json is byte-identical (route names preserved).
* Full suite: 1 failed, 5931 passed, 2 skipped — versus 1 failed, 5886
  passed, 2 skipped on the trunk. The delta is exactly the 45 contract
  goldens added in the previous commit; the one failure is the known
  worktree-environmental test_check_image_formats::TestRepositoryIsClean.
* tests/test_debug_contract.py, test_debug_endpoints.py, test_logs.py and
  test_network_diagnostics_endpoint.py all pass untouched (124 tests).

Two mechanical adaptations, neither of which changes a served byte:

* relative imports inside handler bodies (``from .time_service import ...``)
  became absolute, since the module now sits one level deeper;
* ``debug_get_system_info`` read the module global ``_service_running``
  directly. A call-time proxy cannot stand in for a global read, so it calls
  ``_service_is_running()`` — same flag, same value.

Collaborators still living in api_server are reached through thin call-time
proxies, the src/mqtt/commands.py pattern the other extracted routers use, so
every existing patch("src.api_server.<name>") target stays live. They are
retired later in this same slice. src/debug/routes.py joins the extracted
routers' lint-ignore set in pyproject.toml for the same reason: an extraction
that reformats is not an extraction.
