+++ src/debug/__init__.py
@@ -1,0 +1,7 @@
+"""Debug and diagnostics endpoints.
+
+The board-poking tools (blank / fill / show info), the connection probe, the
+cache inspectors, the system-info card, the network diagnostics runner, and
+the log reader — everything Settings → Debug and the support flow call when
+something is wrong with a board.
+"""
